requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0
supabase==2.3.5

# API Clients
httpx==0.25.1
python-amadeus==8.3.0
aviationstack==1.0.3
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dateutil==2.8.2
pandas-ta==0.3.14b0

# ML
joblib==1.3.2
scikit-learn==1.3.2
xgboost==2.0.2

# Utils
pydantic==2.4.2
pydantic-settings==2.0.3
python-slugify==8.0.1
loguru==0.7.2

# Frontend
jinja2==3.1.2
aiofiles==23.2.1
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
flake8==6.1.0
pylint==3.0.2

# Production
gunicorn==21.2.0
uvloop==0.19.0

# Caching
redis==5.0.1
